backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.3.2
requests==2.31.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
pandas>=2.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
google-auth==2.26.0
google-auth-httplib2==0.1.0